requirements/base.txt
Django==4.1.7
django-environ==0.9.0
psycopg2-binary==2.9.5
uvicorn==0.20.0
djangorestframework==3.14.0